proyecto_rc/proyecto_rc-main/requirements.txt
asgiref==3.9.1
cffi==2.0.0
cryptography==46.0.2
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.1.1
sqlparse==0.5.3
